Add bundle IDs to open, improve app generator
